Add send_from field to notify.mail config
Add new configuration field, notify.mail.send_from, that if set, is used
as the "From:" e-mail field. If not set, the behaviour remains the same,
that smtp_username is used as "From:" too, to be compatible with
previous config files.

This solves the email notification issue where gomail fails to send the
email because the smtp_username is not an e-mail address, e.g.:

  ERRO[0001] Send notification with error:gomail: could not send email
  1: gomail: invalid address "username": mail: missing '@' or angle-addr